requirements.txt
# Core Data Science
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0

# Time Series
statsmodels==0.14.0

# Advanced ML
xgboost==1.7.6
lightgbm==4.0.0

# Deep Learning
tensorflow==2.13.0
keras==2.13.1

# API
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.12.0

# Dashboard
streamlit==1.26.0
plotly==5.16.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2
requests==2.31.0
httpx==0.24.1

# Monitoring & Logging
loguru==0.7.0
prometheus-client==0.17.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.24.1

# Data Validation
great-expectations==0.17.15
pandera==0.17.2
